requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
apify==3.0.1
apify_client==2.2.0
apify_shared==2.1.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
asynctest==0.13.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
colorama==0.4.6
comm==0.2.3
crawlee==1.0.2
cryptography==46.0.2
css-html-js-minify==2.5.5
dateparser==1.2.2
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.21.2
executing==2.2.1
ExifRead==3.5.1
fastjsonschema==2.21.2
feedparser==6.0.12
filelock==3.20.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imagesize==1.4.1
impit==0.7.1
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
langchain-core==0.3.79
langchain-openai==0.3.35
langsmith==0.4.34
lark==1.3.0
lazy-object-proxy==1.12.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
mmh3==5.2.0
more-itertools==10.8.0
multidict==6.7.0
mutagen==1.47.0
narwhals==2.9.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
newsapi==0.1.1
newsapi-python==0.2.7
newsdataapi==0.1.26
nltk==3.9.2
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
openai==2.4.0
orjson==3.11.3
p2j==1.3.2
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
proto-plus==1.26.1
protobuf==5.29.5
proxycheck==0.0.4
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pygooglenews==0.1.3
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
roman-numerals-py==3.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
Send2Trash==1.8.3
setuptools==80.9.0
sgmllib3k==1.0.0
shodan==1.31.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8
Sphinx==8.2.3
sphinx-material==0.0.36
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-trio==1.1.2
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
tiktoken==0.12.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
twilio==9.8.3
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
whois==1.20240129.2
widgetsnbextension==4.0.14
xlsxwriter==3.2.9
yarl==1.22.0
yt-dlp==2025.10.14
zstandard==0.25.0
